+++ CHANGELOG.md
@@ -15,7 +15,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Updates
 
 - Updates the `Icon` component to include variants for `Facebook`, `Instagram`, `Tumblr`, `Twitter` and `YouTube`.
-- Updates the `Icon` component to include variants for `Exit`, `Power` and `Settings`.
+- Updates the `Icon` component to include variants for `Building`, `Exit`, `Locator`, `Power` and `Settings`.
 
 ## 1.0.3 (June 9, 2022)