+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Updates
 
+- Updates the `Logo` component to include a variant for `MLN Color`.
 - Updates the `Icon` component to include variants for `Facebook`, `Instagram`, `Tumblr`, `Twitter` and `YouTube`.
 - Updates the `Icon` component to include variants for `Building`, `Exit`, `Locator`, `Power` and `Settings`.