+++ CHANGELOG.md
@@ -10,6 +10,18 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Adds
 
+- Adds the `useNYPLBreakpoints` hook. This hook internally uses Chakra's `useMediaQuery` hook to get the current responsive media query breakpoint.
+
+### Updates
+
+- Updates the `Logo` component to include a variant for `MLN Color`.
+- Updates the `Icon` component to include variants for `Facebook`, `Instagram`, `Tumblr`, `Twitter` and `YouTube`.
+- Updates the `Icon` component to include variants for `Building`, `Exit`, `Locator`, `Power` and `Settings`.
+
+## 1.0.3 (June 9, 2022)
+
+### Adds
+
 - Adds prop validation for the `TextInput` "number" type for the `min` and `max` props.
 - Adds `min` and `max` prop value validation for the `Slider` component, including in the "range" mode.